+++ src/mako/validator/exceptions/ValidatorException.php
@@ -1,0 +1,18 @@
+<?php
+
+/**
+ * @copyright Frederic G. Østby
+ * @license   http://www.makoframework.com/license
+ */
+
+namespace mako\validator\exceptions;
+
+use RuntimeException;
+
+/**
+ * Validator exception.
+ */
+class ValidatorException extends RuntimeException
+{
+
+}