+++ tests/unit/validator/ValidatorTest.php
@@ -9,9 +9,9 @@ namespace mako\tests\unit\validator;
 
 use mako\i18n\I18n;
 use mako\tests\TestCase;
+use mako\validator\exceptions\ValidationException;
 use mako\validator\rules\I18nAwareInterface;
 use mako\validator\rules\RuleInterface;
-use mako\validator\ValidationException;
 use mako\validator\Validator;
 use Mockery;
 use Throwable;